refactor: extract TrackerExportRetryTypes.swift from TrackerExportSupport.swift
Extracts 3 retry-related types (ExportRetryConfiguration struct,
ExportCreateOutcome enum, TrackerExportRetryContext struct) plus a doc comment
from TrackerExportSupport.swift into a new TrackerExportRetryTypes.swift.

Byte-identical move; SHA verified against origin/main.

Initial extraction hit an over-delete: my sed range 93-129 swallowed the enum
TrackerExportSupport's closing brace at line 93. Corrected to 94-129 which
preserves the enum's structural close.

Reduces TrackerExportSupport.swift from 242 to 205 lines.

Tests: 667 (unchanged).
